\name{HNSCgenomicInstability-package}
\alias{HNSCgenomicInstability-package}
\alias{HNSCgenomicInstability}
\docType{package}
\title{
Head and Neck Squamous Carcinoma scRNA-Seq dataset
}
\description{
This package provides scRNASeq expression data for a subset of 500 cells,
selected at random from GSE103322.
}
\details{
The data object GSE103322 contains two variables:
tmp_matrix: Expression profile matrix for 23014 genes (entrezID) x 500 cells.
metadata_tumor: Logical vector indicating whether each cell has been
classified as tumor by Puram et.al Cell 171:1611 (2017)
}
\author{
Mariano Javier Alvarez <malvarez@darwinhealth.com>
Maintainer: Mariano J. Alvarez <malvarez@darwinhealth.com>
}
\keyword{ package }
