\name{tpm_matrix}
\docType{data}
\alias{tpm_matrix}
\title{HNSC scRNAseq for 500 cells}
\description{
This dataset contains 500 cells selected at random from GSE103322.
}
\usage{tmp_matrix}
\format{A numeric matrix of 23014 features (genes) x 500 cells.}
\references{
Puram SV, Tirosh I, Parikh AS, Patel AP et al. Single-Cell Transcriptomic
Analysis of Primary and Metastatic Tumor Ecosystems in Head and Neck Cancer.
Cell 2017 Dec 14;171(7):1611-1624.e24. PMID: 29198524
}
\keyword{datasets}
