\name{tumor_metadata}
\docType{data}
\alias{tumor_metadata}
\title{Tumor metadata for HNSC scRNAseq 500 cells}
\description{
This dataset contains the metadata for the cells classification for GSE103322.
}
\usage{tumor_metadata}
\format{A logical vector indicating whether each of the 500 cells are classified as tumor}
\references{
Puram SV, Tirosh I, Parikh AS, Patel AP et al. Single-Cell Transcriptomic
Analysis of Primary and Metastatic Tumor Ecosystems in Head and Neck Cancer.
Cell 2017 Dec 14;171(7):1611-1624.e24. PMID: 29198524
}
\keyword{datasets}
